fix style warnings
