+++ AwsWrapperDataProvider/Driver/Plugins/GdbFailover/GlobalDbFailoverMode.cs
@@ -28,42 +28,3 @@ public enum GlobalDbFailoverMode
     OutOfHomeReaderOrWriter,
     AnyReaderOrWriter,
 }
-
-/// <summary>
-/// Extension methods for parsing <see cref="GlobalDbFailoverMode"/> from string values.
-/// </summary>
-public static class GlobalDbFailoverModeExtensions
-{
-    private static readonly Dictionary<string, GlobalDbFailoverMode> NameToValue =
-        new(StringComparer.OrdinalIgnoreCase)
-        {
-            ["strict-writer"] = GlobalDbFailoverMode.StrictWriter,
-            ["strict-home-reader"] = GlobalDbFailoverMode.StrictHomeReader,
-            ["strict-out-of-home-reader"] = GlobalDbFailoverMode.StrictOutOfHomeReader,
-            ["strict-any-reader"] = GlobalDbFailoverMode.StrictAnyReader,
-            ["home-reader-or-writer"] = GlobalDbFailoverMode.HomeReaderOrWriter,
-            ["out-of-home-reader-or-writer"] = GlobalDbFailoverMode.OutOfHomeReaderOrWriter,
-            ["any-reader-or-writer"] = GlobalDbFailoverMode.AnyReaderOrWriter,
-        };
-
-    /// <summary>
-    /// Parses a kebab-case string value into a <see cref="GlobalDbFailoverMode"/>.
-    /// </summary>
-    /// <param name="value">The kebab-case string to parse (e.g., "strict-writer"), or null.</param>
-    /// <returns>The corresponding <see cref="GlobalDbFailoverMode"/>, or null if the input is null.</returns>
-    /// <exception cref="ArgumentException">Thrown when the value does not match any known failover mode.</exception>
-    public static GlobalDbFailoverMode? FromValue(string? value)
-    {
-        if (value == null)
-        {
-            return null;
-        }
-
-        if (NameToValue.TryGetValue(value, out var mode))
-        {
-            return mode;
-        }
-
-        throw new ArgumentException($"Invalid GlobalDbFailoverMode value: {value}");
-    }
-}
